requirements.txt
google-genai
flask
